update v 2.0

+++ src/isotope.c
@@ -25,8 +25,8 @@ int set_isotope(Isotope* isotope, char* symbol, char* iso_name, double mass, dou
 
 int isotope_sort_by_abundance(const void *a, const void *b)
 {
-    double y1 = ((const struct Isotope*)a)->abundance;
-    double y2 = ((const struct Isotope*)b)->abundance;
+    long double y1 = ((const struct Isotope*)a)->abundance;
+    long double y2 = ((const struct Isotope*)b)->abundance;
     
     if (y1 < y2) {
         return 1;
@@ -39,11 +39,11 @@ int isotope_sort_by_abundance(const void *a, const void *b)
 
 int isotope2_sort_by_n_abundance_inc(const void *a, const void *b)
 {
-    double x1 = ((const struct Isotope2*)a)->amount;
-    double x2 = ((const struct Isotope2*)b)->amount;
+    size_t x1 = ((const struct Isotope2*)a)->amount;
+    size_t x2 = ((const struct Isotope2*)b)->amount;
     
-    double y1 = ((const struct Isotope2*)a)->abundance;
-    double y2 = ((const struct Isotope2*)b)->abundance;
+    long double y1 = ((const struct Isotope2*)a)->abundance;
+    long double y2 = ((const struct Isotope2*)b)->abundance;
     
     if (x1*y1 > x2*y2) {
         return 1;
@@ -57,11 +57,11 @@ int isotope2_sort_by_n_abundance_inc(const void *a, const void *b)
 
 int isotope2_sort_by_n_abundance_dec(const void *a, const void *b)
 {
-    double x1 = ((const struct Isotope2*)a)->amount;
-    double x2 = ((const struct Isotope2*)b)->amount;
+    size_t x1 = ((const struct Isotope2*)a)->amount;
+    size_t x2 = ((const struct Isotope2*)b)->amount;
     
-    double y1 = ((const struct Isotope2*)a)->abundance;
-    double y2 = ((const struct Isotope2*)b)->abundance;
+    long double y1 = ((const struct Isotope2*)a)->abundance;
+    long double y2 = ((const struct Isotope2*)b)->abundance;
     
     if (x1*y1 < x2*y2) {
         return 1;
